git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@2023 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ Magick++/lib/Geometry.cpp
@@ -210,7 +210,7 @@ Magick::Geometry::operator = ( const std::string &geometry_ )
   ssize_t y = 0;
   size_t width_val = 0;
   size_t height_val = 0;
-  int flags = GetGeometry (geom, &x, &y, &width_val, &height_val );
+  ssize_t flags = GetGeometry (geom, &x, &y, &width_val, &height_val );
 
   if (flags == NoValue)
     {
